[Silver IV] Title: 수들의 합 2, Time: 324 ms, Memory: 30416 KB -BaekjoonHub

+++ 백준/Silver/2003. 수들의 합 2/수들의 합 2.java
@@ -22,7 +22,7 @@ public class Main {
 
         Scanner kb = new Scanner(System.in);
         int n = kb.nextInt();
-        long m = kb.nextInt();
+        int m = kb.nextInt();
 
         int[] arr = new int[n];
